requirements.txt
streamlit
pandas
playwright
openpyxl
gspread
google-auth
